Fixed issue with deleting the zone on close
- Fixed issue with deleting the zone on close. Don't try to delete the zone file if a FF wasn't opened

+++ Call of Duty FastFile Editor/UI/MainWindowForm.cs
@@ -1363,7 +1363,7 @@ namespace Call_of_Duty_FastFile_Editor
                 {
                     e.Cancel = true;
                 }
-                else
+                else if(_openedFastFile != null)
                     File.Delete(_openedFastFile.ZoneFilePath);
             }
         }
